@@ -1,4 +1,3 @@
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:url_strategy/url_strategy.dart';
@@ -6,14 +5,13 @@ import 'views/view/enter_code.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp();
   setPathUrlStrategy();
 
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -24,22 +22,13 @@ class MyApp extends StatelessWidget {
       ),
       onGenerateRoute: (settings) {
         // https://stackoverflow.com/a/59755970/13617136
-        List<String> pathComponents = settings.name.split('/');
-        switch (settings.name) {
-          case '/':
-            return MaterialPageRoute(
-              builder: (context) => const EnterCode(),
-            );
-            break;
-          default:
-            return MaterialPageRoute(
-              builder: (context) => EnterCode(
-                pathComponents.last,
-              ),
-            );
-        }
+        List<String> pathComponents = settings.name!.split('/');
+        return MaterialPageRoute(
+          builder: (context) => EnterCode(
+            userCode: pathComponents.last,
+          ),
+        );
       },
-      home: const EnterCode(),
     );
   }
 }
