@@ -11,8 +11,8 @@ class CustomSnack {
   }
 
   static void showSnack(BuildContext context,
-      {@required String message,
-      SnackBarAction barAction,
+      {required String message,
+      SnackBarAction? barAction,
       Duration duration = const Duration(milliseconds: 2100)}) {
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       action: barAction,
