@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 
 class LoadingIndicator extends StatelessWidget {
   const LoadingIndicator({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
